Yet Another Nav Tag
Added ‘navitem’ tag to allow setting the ‘active’ class on the right
nav item(s). Each nav item has a slug. To set the active state of the
nav a view returns a ‘nav’ context var which is a list of nav slugs.
Each nav item with a matching slug is activated.